refactor(raijin): group project owners
