+++ yarn/raijin/src/project/index.ts
@@ -1,4 +1,4 @@
-export *      from './manifest.js'
+export *      from './manifest/manifest.js'
 export type * from './project.interfaces.js'
-export *      from './type.js'
-export *      from './workspaces.js'
+export *      from './type/type.js'
+export *      from './workspaces/workspaces.js'
