+++ yarn/raijin/src/project/manifest/manifest.ts
@@ -1,4 +1,4 @@
-import type { ProjectManifestSource } from './project.interfaces.js'
+import type { ProjectManifestSource } from '../project.interfaces.js'
 
 export const getManifestWorkspacePatterns = (manifest: ProjectManifestSource): Array<string> =>
   manifest.workspaceDefinitions.map(({ pattern }) => pattern)
