+++ yarn/raijin/src/project/workspaces/workspaces.test.ts
@@ -24,52 +24,16 @@ const createProject = (workspaces: Array<Workspace>): Project =>
     workspaces,
   }) as Project
 
-test('should resolve single project model without workspace patterns', () => {
-  const rootWorkspace = createWorkspace('/repo', createManifest({}))
-  const model = createProjectModel(createProject([rootWorkspace]))
-
-  assert.equal(model.cwd, '/repo')
-  assert.equal(model.topLevelWorkspace, rootWorkspace)
-  assert.equal(model.type, 'single')
-  assert.deepEqual(model.workspacePatterns, [])
-  assert.deepEqual(model.workspaces, [rootWorkspace])
-})
-
-test('should resolve monorepo project model from array workspace patterns', () => {
+test('should project the top-level workspace and normalized project state', () => {
   const rootWorkspace = createWorkspace('/repo', createManifest({ workspaces: ['packages/*'] }))
   const clientWorkspace = createWorkspace('/repo/packages/client', createManifest({}))
   const model = createProjectModel(createProject([rootWorkspace, clientWorkspace]))
 
+  assert.equal(model.cwd, rootWorkspace.cwd)
+  assert.equal(model.topLevelWorkspace, rootWorkspace)
   assert.equal(model.type, 'monorepo')
   assert.deepEqual(model.workspacePatterns, ['packages/*'])
   assert.deepEqual(model.workspaces, [rootWorkspace, clientWorkspace])
-})
-
-test('should resolve monorepo project model from object workspace patterns', () => {
-  const rootWorkspace = createWorkspace(
-    '/repo',
-    createManifest({
-      workspaces: {
-        packages: ['apps/*', 'packages/*'],
-      },
-    })
-  )
-  const model = createProjectModel(createProject([rootWorkspace]))
-
-  assert.equal(model.type, 'monorepo')
-  assert.deepEqual(model.workspacePatterns, ['apps/*', 'packages/*'])
-})
-
-test('should follow normalized Yarn workspace definitions', () => {
-  const manifest = createManifest()
-
-  manifest.workspaceDefinitions.push({ pattern: 'packages/*' })
-
-  const rootWorkspace = createWorkspace('/repo', manifest)
-  const model = createProjectModel(createProject([rootWorkspace]))
-
-  assert.equal(model.type, 'monorepo')
-  assert.deepEqual(model.workspacePatterns, ['packages/*'])
 })
 
 test('should keep the project boundary at the root when a leaf declares Raijin', () => {
