+++ yarn/raijin/src/project/workspaces/workspaces.ts
@@ -1,9 +1,9 @@
-import type { ProjectSource }           from './project.interfaces.js'
-import type { ProjectWorkspaceSource }  from './project.interfaces.js'
-import type { RaijinProjectModel }      from './project.interfaces.js'
+import type { ProjectSource }           from '../project.interfaces.js'
+import type { ProjectWorkspaceSource }  from '../project.interfaces.js'
+import type { RaijinProjectModel }      from '../project.interfaces.js'
 
-import { getManifestWorkspacePatterns } from './manifest.js'
-import { resolveProjectType }           from './type.js'
+import { getManifestWorkspacePatterns } from '../manifest/manifest.js'
+import { resolveProjectType }           from '../type/type.js'
 
 export const createProjectModel = <TWorkspace extends ProjectWorkspaceSource>(
   project: ProjectSource<TWorkspace>
